+++ Classes/Mapper/BackEndUserMapper.php
@@ -4,7 +4,6 @@ declare(strict_types=1);
 
 namespace OliverKlee\Oelib\Mapper;
 
-use OliverKlee\Oelib\Exception\NotFoundException;
 use OliverKlee\Oelib\Model\BackEndUser;
 
 /**
@@ -15,29 +14,4 @@ class BackEndUserMapper extends AbstractDataMapper
     protected $tableName = 'be_users';
 
     protected $modelClassName = BackEndUser::class;
-
-    /**
-     * @deprecated #1505 will be removed in oelib 6.0.0
-     */
-    protected $additionalKeys = ['username'];
-
-    /**
-     * Finds a back-end user by username. Hidden user records will be retrieved
-     * as well.
-     *
-     * @param non-empty-string $username username, case-insensitive
-     *
-     * @return BackEndUser model of the back-end user with the provided username
-     *
-     * @throws NotFoundException if there is no back-end user with the provided username in the be_user table
-     *
-     * @deprecated #1505 will be removed in oelib 6.0.0
-     */
-    public function findByUserName(string $username): BackEndUser
-    {
-        /** @var BackEndUser $result */
-        $result = $this->findOneByKey('username', $username);
-
-        return $result;
-    }
 }