+++ Tests/Functional/Mapper/BackEndUserMapperTest.php
@@ -4,9 +4,7 @@ declare(strict_types=1);
 
 namespace OliverKlee\Oelib\Tests\Functional\Mapper;
 
-use OliverKlee\Oelib\Exception\NotFoundException;
 use OliverKlee\Oelib\Mapper\BackEndUserMapper;
-use OliverKlee\Oelib\Model\BackEndUser;
 use TYPO3\TestingFramework\Core\Functional\FunctionalTestCase;
 
 /**
@@ -40,73 +38,5 @@ final class BackEndUserMapperTest extends FunctionalTestCase
         $this->subject->load($model);
 
         self::assertSame('max', $model->getUserName());
-    }
-
-    // Tests concerning findByUserName
-
-    /**
-     * @test
-     */
-    public function findByUserNameWithNameOfExistingUserReturnsMatchingBackEndUserInstance(): void
-    {
-        $this->importDataSet(__DIR__ . '/Fixtures/BackEndUsers.xml');
-
-        $model = $this->subject->findByUserName('max');
-
-        self::assertInstanceOf(BackEndUser::class, $model);
-        self::assertSame('max', $model->getUserName());
-    }
-
-    /**
-     * @test
-     */
-    public function findByUserNameWithUppercasedNameOfExistingLowercasedUserReturnsMatchingModel(): void
-    {
-        $this->importDataSet(__DIR__ . '/Fixtures/BackEndUsers.xml');
-
-        $model = $this->subject->findByUserName('MAX');
-
-        self::assertInstanceOf(BackEndUser::class, $model);
-        self::assertSame('max', $model->getUserName());
-    }
-
-    /**
-     * @test
-     */
-    public function findByUserNameWithUppercasedNameOfExistingUppercasedUserReturnsMatchingModel(): void
-    {
-        $this->importDataSet(__DIR__ . '/Fixtures/BackEndUsers.xml');
-
-        $model = $this->subject->findByUserName('MOE');
-
-        self::assertInstanceOf(BackEndUser::class, $model);
-        self::assertSame('MOE', $model->getUserName());
-    }
-
-    /**
-     * @test
-     */
-    public function findByUserNameWithLowercaseNameOfExistingUppercaseUserReturnsModelWithThatUid(): void
-    {
-        $this->importDataSet(__DIR__ . '/Fixtures/BackEndUsers.xml');
-
-        $model = $this->subject->findByUserName('moe');
-
-        self::assertInstanceOf(BackEndUser::class, $model);
-        self::assertSame('MOE', $model->getUserName());
-    }
-
-    /**
-     * @test
-     */
-    public function findByUserNameNotFindsDeletedUser(): void
-    {
-        $this->importDataSet(__DIR__ . '/Fixtures/BackEndUsers.xml');
-
-        $this->expectException(NotFoundException::class);
-        $this->expectExceptionMessage('No records found in the table "be_users" matching: {"username":"deleted"}');
-        $this->expectExceptionCode(0);
-
-        $this->subject->findByUserName('deleted');
     }
 }